feat(lib/AWSstub): exposed the mocked node aws client
Docs(ReadMe): added info about requiring AWSstub

docs(ReadMe): fixed spelling error
